requirements.txt
numpy==1.16.3
Pillow==6.0.0
torch==1.1.0
torchvision==0.2.2
tensorflow==1.13.1
tensorboard==1.13.1
scipy==1.1.0
scikit-learn==0.21.1
requests==2.22.0
opencv-contrib-python==4.1.0.25
h5py==2.9.0
tqdm==4.32.1
imutils==0.5.2